bb/requirements.txt
# Bennett Bites — Python dependencies
# Install with:  pip install -r requirements.txt

fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy==2.0.30

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# File uploads (multipart form-data parsing)
python-multipart==0.0.9

# Pydantic email validation
pydantic[email]==2.7.1
